fix(state-home): extract projectKey to a light module to break import coupling (PAN-2372)
The full test suite (run as the final pre-done gate) caught two regressions
introduced by WI-1's resolveStateReadHomeSync change — both missed by the
per-bead targeted test runs. They share one root cause.

Root cause: WI-1 made resolveStateReadHomeSync (state-read-home.ts) import
projectKey from state-home.ts. state-home.ts is a heavyweight module — it
imports child_process (execFile, for async git ops) and state-plane.ts (which
also accesses child_process at module load). That transitive coupling flowed
into the lightweight sync-read door and broke unrelated tests:

- compact-beads.test.ts / pan-639-compact-beads-persistence.test.ts: these
  partially mock child_process without execFile. Loading state-read-home.ts now
  transitively evaluated state-plane.ts's `promisify(execFile)` at import time,
  crashing the suite before any test ran.
- done.test.ts: resolveStateReadHomeSync now reached listProjectsSync() (via
  projectKey), but the test's partial projects.js mock did not export it.

Fix: move projectKey to a dedicated light module, src/lib/project-key.ts, that
imports only listProjectsSync + ProjectConfig from projects.js (clean of
child_process / state-plane) plus node:path. Both the async door (state-home.ts)
and the sync door (state-read-home.ts) import projectKey from there — single
implementation preserved (the state-read-home parity test still holds), and
state-read-home.ts no longer drags in child_process. project-key.ts imports
listProjectsSync from projects.js, so a vi.mock of projects.js still steers the
registry (the parity test's mock strategy is intact).

Companion: done.test.ts's projects.js mock gains listProjectsSync: () => [] so
projectKey falls back to the path basename; the done preflight/shadow tests do
not assert on the resolved state worktree.

Verified: full `npm test` green (971 passed | 4 skipped), typecheck + lint pass.
Baseline: none of the touched files are file-size baselined; project-key.ts is a
new ~30-line module.
